+++ user-rest-service/interfaces/handler/common.go
