+++ user-rest-service/interfaces/handler/common_test.go
