+++ user-rest-service/interfaces/handler/group.go
