+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_DeleteGroupApprovedUser/response.json.golden
