+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_DeleteGroupUnapprovedUser/response.json.golden
