+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_GetGroupList/response.json.golden
