+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_GetGroupUserIDList/response.json.golden
