+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_GetUser/response.json.golden
