+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_Logout/response.json.golden
