+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PostGroup/request.json
