+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PostGroup/response.json.golden
