+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PostGroupApprovedUser/response.json.golden
