+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PostGroupUnapprovedUser/request.json
