+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PostGroupUnapprovedUser/response.json.golden
