+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PutGroup/request.json
