+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_PutGroup/response.json.golden
