+++ user-rest-service/interfaces/handler/testdata/TestDBHandler_VerifyGroupAffiliationOfUsersList/request.json
