+++ user-rest-service/interfaces/handler/testdata/Test_userHandler_Login/response.json.golden
