+++ user-rest-service/interfaces/handler/testdata/Test_userHandler_SignUp/response.json.golden
