+++ user-rest-service/interfaces/presenter/user_errors.go
@@ -1,4 +1,4 @@
-package userdomain
+package presenter
 
 import (
 	"encoding/json"
